web: * web初始化失败后自动回滚操作 * 检查数据线bug修复 * 修复kafka reader页面卡顿的bug * zookeeper部分模板修改 * 拓扑stopped下，表任意状态都可以删除 * 多租户namespace修改,在instance后添加router名称用!分隔 * 资源管理tab页，已选择列表带过滤功能 * 优化资源添加:添加一键添加所有表功能 * 全量拉取记录添加编辑功能 * 一键清除全量报警(数据线级别) * 登录界面添加账户名为邮箱的提示，防止误解 全量: * 全量拉取主键为bigint的表数据条数对不上的bug修复(bugbigint转成int丢失精度导致) * 全量mysql脱敏列大小写bug修复(查询语句列名转成了大写)
* canal-auto-deploy  bugfix

+++ dbus-keeper/keeper-mgr/src/main/java/com/creditease/dbus/controller/ProjectController.java
@@ -51,6 +51,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import static com.creditease.dbus.constant.MessageCode.PROJECT_NOT_ALLOWED_DELETED;
+
 /**
  * Created by zhangyf on 2018/3/7.
  *
@@ -208,6 +210,10 @@ public class ProjectController extends BaseController {
     @GetMapping("/delete/{id}")
     @ProjectAuthority
     public ResultEntity deleteById(@PathVariable("id") int id) {
+        //判断能否直接删除project
+        if(service.getRunningTopoTables(id) != 0){
+            return resultEntityBuilder().status(PROJECT_NOT_ALLOWED_DELETED).build();
+        }
         return service.deleteProject(id);
     }
 
@@ -234,6 +240,11 @@ public class ProjectController extends BaseController {
     @GetMapping(path = "search")
     public ResultEntity search(HttpServletRequest request) {
         return service.search(request.getQueryString());
+    }
+
+    @GetMapping("/getAllResourcesByQuery")
+    public ResultEntity getAllResourcesByQuery(HttpServletRequest request) throws Exception{
+        return service.getAllResourcesByQuery(request.getQueryString());
     }
 
 }
